+++ src/gas/attributes/attribute_snapshot.rs
@@ -1,19 +1,19 @@
 #[derive(Debug, Clone, Copy)]
 pub struct AttributeSnapshot {
-    base: f64,
-    current: f64,
+    base: f32,
+    current: f32,
 }
 
 impl AttributeSnapshot {
-    pub fn new(base: f64, current: f64) -> Self {
+    pub fn new(base: f32, current: f32) -> Self {
         Self { base, current }
     }
 
-    pub fn base(&self) -> f64 {
+    pub fn base(&self) -> f32 {
         self.base
     }
 
-    pub fn current(&self) -> f64 {
+    pub fn current(&self) -> f32 {
         self.current
     }
 }
